requirements.txt
python-binance
TA-lib
json
numpy
Flask